Renombrando archivos para consistencia
Este cambio hace que dos archivos de `/admin/` tengan un nombre
consistente con su URL.
