+++ frontend/www/arena/admin.php
@@ -6,12 +6,11 @@ require_once(dirname(__DIR__, 2) . '/server/bootstrap.php');
 \OmegaUp\UITools::render(
     function (\OmegaUp\Request $r): array {
         return [
-            'smartyProperties' => [],
-            'template' => 'arena.contest.admin.tpl',
-            'inContest' => (
-                !isset($_REQUEST['is_practice']) ||
-                boolval($_REQUEST['is_practice']) !== true
-            ),
+            'smartyProperties' => [
+                'title' => 'wordsGlobalSubmissions',
+                'fullWidth' => true,
+            ],
+            'entrypoint' => 'arena_admin',
         ];
     }
 );
